hrt_ioctl: Move common code under /common folder
The HRT ioctl is common code, so move it out of chip specific code
